Conflicts resolved:
- csrc/libtorch_stable/fp32_router_gemm.cu, tests/kernels/test_fp32_router_gemm.py:
  formatting-only (upstream and this branch's geometry identical); kept this
  branch's side.
- vllm/.../router/gate_linear.py: kept upstream's superset docstring adding the
  cuteDSL ll_bf16_gemm dispatch tier; forward() retains all five tiers.

+++ vllm/model_executor/models/hy_v3.py
@@ -177,7 +177,9 @@ class HYV3MoEFused(nn.Module):
         else:
             self.shared_mlp = None
 
-        self.expert_bias = nn.Parameter(torch.empty(config.num_experts))
+        self.expert_bias = nn.Parameter(
+            torch.empty(config.num_experts, dtype=torch.float32)
+        )
         scoring_func = "sigmoid"
         e_score_correction_bias = self.expert_bias
